requirements.txt
Pillow
Rich
#Keyboard (requires root)
#Pynput (requires X)
#getch (pip error)
